feature:implementing i18n with next-intl. Error: Cannot read properties of undefined (reading 'toString')
