+++ app/[locale]/(Dashboard)/courses/loading.jsx
