+++ app/[locale]/(Dashboard)/layout.jsx
