+++ app/[locale]/(Dashboard)/saved/page.jsx
@@ -4,17 +4,16 @@ import Link from 'next/link'
 import { useContext } from 'react'
 import { IoIosStats } from "react-icons/io"
 import { useEffect, useState } from "react"
-import { useSelector, useDispatch  } from "react-redux";
-
+import { useSelector, useDispatch } from "react-redux";
+import { setRecommendedData } from '@/store/userSlice';
+import { Reveal } from '@/app/[locale]/components/UI/Reveal';
 
 
 
 import { DarkModeContext } from '../layout'
-import CourseDescription from '@/app/components/courses/CourseDescription'
-import CourseCard from '@/app/components/courses/CourseCard'
-import getCourses from "@/app/lib/getCourses"
-import { setRecommendedData } from '@/store/userSlice';
-import { Reveal } from '@/app/components/UI/Reveal'
+import CourseDescription from '@/app/[locale]/components/courses/CourseDescription'
+import CourseCard from '@/app/[locale]/components/courses/CourseCard'
+import getCourses from "@/app/[locale]/lib/getCourses"
 
 
 
@@ -22,15 +21,14 @@ const page = () => {
   const [courses, setCourses] = useState([]);
   const [courseId, setCourseId] = useState('09cb9789-12c7-4c4a-9513-fa76146d0017')
   const [selectedCourse, setSelectedCourse] = useState(courses[0]);
-  const user = useSelector((state) => state.user);
   const dispatch = useDispatch(); 
-
+  const user = useSelector((state) => state.user);
   
 // console.log(courses)
 
 
   
-  const savedCoursesId = user.myCourses.map((obj) => obj.id)
+  const savedCoursesId = user.savedCourses.map((obj) => obj.id)
 
   const filteredCourses = courses.filter((course) =>
   savedCoursesId.includes(course.id)
@@ -41,9 +39,6 @@ const page = () => {
   const handleSendData = (recommendedData) => {
     dispatch(setRecommendedData(recommendedData));
   };
-
-
-
 
   useEffect(() => {
     const fetchData = async () => {
@@ -110,38 +105,42 @@ const page = () => {
     : 
     
     (<div className={`${style.main} ${isDark ? dark.main : light.main}`}>
-      <div className={style.left_section}>
+     <div className={style.left_section}>
       <Reveal>
         <div className='flex justify-between'>
-          <h1 className='text-3xl'>My Learning</h1>
+          <h1 className='text-3xl'>Saved Courses</h1>
           <div className='flex flex-row items-center space-x-1'>
             <IoIosStats className='text-stone-500' />
             <p className='font-normal text-stone-500'>Statistics</p>
           </div>
         </div>
-      </Reveal>
-
+      </Reveal>  
+      
+      
       <Reveal delay="0.6">
         <div className='space-y-5'>
          {filteredCourses.map(course => (
             <CourseCard key={course.id} title={course.courseName} trainer={course.trainerName} imgUrl={course.img} isDark={isDark} courseId={course.id}
               setCourseId={setCourseId} />
-          ))}
-              </div>
-      </Reveal>      
-            </div>
+              ))}
+        </div>
+      </Reveal>    
+      </div>
       <Reveal delay="0.6" >
-            <div className={`${style.right_section} ${isDark ? dark.right_section : light.right_section}`}>
-              <CourseDescription isDark={isDark} selectedCourse={selectedCourse} />
-              <div className={style.buttonContainer}>
-                <button onClick={() => handleSendData({courseId: selectedCourse.id })} className={`${style.button} ${isDark ? dark.button : light.button}`}>REVIEW COURSE</button>
-                <Link href="/#">
-                  <button className={`${style.linkButton} ${isDark ? dark.linkButton : light.linkButton}`}>CONTINUE LEARNING</button>
-                </Link>
-              </div>
-            </div>
-      </Reveal>        
-          </div>)
+      <div className={`${style.right_section} ${isDark ? dark.right_section : light.right_section}`}>
+        <CourseDescription isDark={isDark} selectedCourse={selectedCourse}/>
+        <div className={style.buttonContainer}>
+        <Link href="./overview">
+          <button onClick={() => handleSendData({courseId: selectedCourse.id })} className={`${style.button} ${isDark ? dark.button : light.button}`}>REVIEW COURSE</button>
+        </Link>
+          <Link href="/#">
+            <button className={`${style.linkButton} ${isDark ? dark.linkButton : light.linkButton}`}>CONTINUE LEARNING</button>
+          </Link>
+        </div>
+      </div>
+      </Reveal>  
+     
+    </div>)
           
   )
 }
