+++ app/[locale]/components/UI/Reveal.jsx
