+++ app/[locale]/components/courses/CourseCard.jsx
