+++ app/[locale]/components/courses/CourseDescription.jsx
