+++ app/[locale]/components/home/CategoryCard.jsx
