+++ app/[locale]/components/home/CoursesCard.jsx
