+++ app/[locale]/components/home/LearningCard.jsx
