+++ app/[locale]/components/profile/Info.jsx
