+++ app/[locale]/components/profile/Statistics.jsx
