+++ app/[locale]/components/search/RatingFilter.jsx
