+++ app/[locale]/favicon.ico
