+++ app/[locale]/firebase.js
