+++ app/[locale]/globals.css
