+++ app/[locale]/layout.jsx
@@ -2,13 +2,8 @@ import ProviderWrapper from '@/store/provider'
 import { UserProvider } from "@auth0/nextjs-auth0/client";
 import './globals.css'
 import { Inter } from 'next/font/google'
-
-
-// import {useLocale} from 'next-intl';
+import {useLocale} from 'next-intl';
 import {notFound} from 'next/navigation';
-// import {NextIntlClientProvider} from 'next-intl';
-
-
 
  
 const inter = Inter({ subsets: ["latin"] })
@@ -62,26 +57,20 @@ export const metadata = {
   ]
 }
 export default async function RootLayout({ children,params }) {
-  // const locale = useLocale();
-
-  // let messages;
-  // try {
-  //   messages = (await import(`../../messages/${locale}.json`)).default;
-  // } catch (error) {
-  //   notFound();
-  // }
+  const locale = useLocale();
+ 
+  // Show a 404 error if the user requests an unknown locale
+  if (params.locale !== locale) {
+    notFound();
+  }
 
   return (
-    <html lang="en">
-      {/* <NextIntlClientProvider locale={locale} messages={messages}> */}
-
+    <html lang={locale}>
       <ProviderWrapper>
         <UserProvider>
           <body className={inter.className}>{children}</body>
         </UserProvider>
       </ProviderWrapper>
-
-      {/* </NextIntlClientProvider> */}
     </html>
   )
 }
