+++ app/[locale]/lib/getCourses.js
