+++ app/[locale]/loading.jsx
