+++ middleware.js
@@ -1,73 +1,32 @@
-import { withMiddlewareAuthRequired } from '@auth0/nextjs-auth0/edge';
+// import { withMiddlewareAuthRequired } from '@auth0/nextjs-auth0/edge';
 
-export default withMiddlewareAuthRequired();
+// export const intlMiddlewareAuth = withMiddlewareAuthRequired();
 
-export const config = {
-  matcher: ['/profile', '/courses',"/saved"]
-};
-
+// export const config = {
+//   matcher: ['/profile', '/courses',"/saved",'/((?!api|_next|.*\\..*).*)']
+// };
 
 // import createMiddleware from 'next-intl/middleware';
  
-// export default intlMiddleware =  createMiddleware({
+// export const intlMiddleware =  createMiddleware({
 //   // A list of all locales that are supported
-//   locales: ['en', 'tr'],
+//   locales: ['en', 'de'],
  
 //   // If this locale is matched, pathnames work without a prefix (e.g. `/about`)
-//   defaultLocale: 'en',
-//   localePrefix: 'never',
+//   defaultLocale: 'en'
 // });
+import createMiddleware from 'next-intl/middleware';
  
-// export const config = {
-//   // Skip all paths that should not be internationalized. This example skips the
-//   // folders "api", "_next" and all files with an extension (e.g. favicon.ico)
-//   matcher: ['/((?!api|_next|.*\\..*).*)']
-// };
-
-
-// import { withAuth } from 'next-auth/middleware';
-// import createIntlMiddleware from 'next-intl/middleware';
-
-// const locales = ['en', 'de'];
-// const publicPages = ['/', '/login', '/home', '/overview', '/search'];
-
-// const intlMiddleware = createIntlMiddleware({
-//   locales,
-//   defaultLocale: 'en',
-//   localePrefix: 'never',
-// });
-
-// const authMiddleware = withAuth(
-//   // Note that this callback is only invoked if
-//   // the authorized callback has returned true
-//   // and not for pages listed in pages.
-//   function onSuccess(req) {
-//     return intlMiddleware(req);
-//   },
-//   {
-//     callbacks: {
-//       authorized: ({ token }) => token != null,
-//     },
-//     pages: {
-//       signIn: '/login',
-//     },
-//   }
-// );
-
-// export default function middleware(req) {
-//   const publicPathnameRegex = new RegExp(
-//     '^(/(' + locales.join('|') + '))?(' + publicPages.join('|') + ')?/?$',
-//     'i'
-//   );
-//   const isPublicPage = publicPathnameRegex.test(req.nextUrl.pathname);
-
-//   if (isPublicPage) {
-//     return intlMiddleware(req);
-//   } else {
-//     return authMiddleware(req);
-//   }
-// }
-
-// export const config = {
-//   matcher: ['/((?!api|_next|.*\\..*).*)'],
-// };
+export default createMiddleware({
+  // A list of all locales that are supported
+  locales: ['en', 'de'],
+ 
+  // If this locale is matched, pathnames work without a prefix (e.g. `/about`)
+  defaultLocale: 'en'
+});
+ 
+export const config = {
+  // Skip all paths that should not be internationalized. This example skips the
+  // folders "api", "_next" and all files with an extension (e.g. favicon.ico)
+  matcher: ['/((?!api|_next|.*\\..*).*)']
+};
