+++ next.config.js
@@ -18,5 +18,9 @@ const nextConfig = {
   },
 };
 
+const withNextIntl = require('next-intl/plugin')(
+  './i18n.js'
+);
 
-module.exports = nextConfig;
+
+module.exports = withNextIntl(nextConfig);
