+++ package.json
@@ -18,7 +18,7 @@
     "eslint-config-next": "13.4.7",
     "firebase": "^9.23.0",
     "framer-motion": "^10.15.1",
-    "next": "^13.4.7",
+    "next": "^13.4.13",
     "next-intl": "^3.0.0-beta.9",
     "postcss": "8.4.24",
     "react": "18.2.0",
